# feat & refactor : Issue PostPage 제작 및 css, 컴포넌트 정리
1. Issue PostPage 제작
- Issue Card 형태와 Post 형태의 컴포넌트를 합쳐서 구현함
- 지금은 true, false로 구분하고 있지만 상수화할때 상수로 변경할 예정
- Issue Card의 drag 불특정하게 작동 (Container안에 Container가 들어가거나, Card가 Container 옆으로 나오는 현상)

2. css, 파일 정리
- IssueList, IssueItem를 Issue 폴더로 이동
- IssueCard의 mark에 className을 다시 변경 (css 미적용 해결)
- ojbectivePopup css, popup.scss로 이동
- IssueItem을 IssuePostItem으로 명명 (불분명한 네이밍)
