+++ src/pages/Project/MainView/index.tsx
@@ -2,18 +2,18 @@ import React from 'react'
 
 import ProjectBoard from './ProjectBoard'
 import FeedList from './FeedList'
-import IssueList from './IssueList'
+import IssueList from '../Issue/IssueList'
 import ProjectDescription from './Description'
 
-const MainView = () => {
-
+const MainView = (props) => {
+    const {projectId} = props
     return (
         <div className="MainView">
             <div className="publishFile">
                 <ProjectBoard/>
                 <FeedList/>
             </div>
-            <IssueList/>
+            <IssueList projectId={projectId}/>
             <ProjectDescription/>
         </div>
     )
