+++ src/scss/component/issuelist.scss
@@ -1,0 +1,173 @@
+.issueList {
+    & > .head {
+        @include flexbox(space-between, center);
+        @include commonTitleStyle();
+    
+        & .left button {
+            font-size: 18px;
+            color: #928a76;
+            margin: 0 15px;
+        }
+        
+        & .right {
+            @include flexbox(space-between, center);
+
+            & > * {
+                margin-left: 20px;
+            }
+        }
+
+        & .right .moreTag {
+            background-repeat: no-repeat;
+            background-position: right;
+            border: 1px solid #717f90;
+            font-size: 18px;
+            height: 34px;
+            border-radius: 5px;
+            background-color: #f2f2f2;
+            padding: 0 15px;
+            margin-right: 30px;
+        }
+
+       
+    }
+}
+
+.issueListBody {
+    width: 100%;
+
+    &.borderTop {
+        border-top: 1px solid $ColorNavy;
+    }
+    
+    & .item {
+        text-align: center;
+        @include flexbox(space-between, center);
+        padding: 15px;
+        border-bottom: 1px solid $ColorNavy;
+    
+        & > * {
+            margin: 0 5px;
+        }
+    }
+    
+    & .iconbox {
+        margin-right: 40px;
+    }
+    
+    & .textbox {
+        @include notCutToNextLine();
+        display: inline-block;
+    
+        width: 100%;
+        text-align: left;
+        & .text1 {
+            overflow: hidden;
+            text-overflow: ellipsis;
+            & .skillMark {
+                margin-left: 10px;
+            }
+    
+            & span {
+                font-size: 18px;
+                max-width: 100%;
+            }
+        }
+    
+        & .text2 {
+            font-size: 14px;
+            color: #888;
+            padding-left: 20px;
+            overflow: hidden;
+            text-overflow: ellipsis;
+        }
+    }
+    
+    & .num1box {
+        width: 100px;
+    }
+    & .num2box {
+        width: 100px;
+    }
+    & .gradebox {
+        & img {
+            margin-bottom: 5px;
+        }
+        & p {
+            font-size: 14px;
+        }
+    }
+    
+    & .userbox {
+        width: 150px;
+        white-space : pre;
+        & p {
+            color: #707070;
+            font-size: 12px;
+        }
+        & img {
+            margin-top: 15px;
+        }
+    }
+}
+
+@media screen and (max-width: 780px) {
+    .issueList {
+        & .head {
+            & .right .search {
+                width: 150px;
+            }
+        
+            & .right > * {
+                margin-left: 10px;
+            }
+        
+            & .left .option {
+                @include flexbox(space-between, center);
+            }
+
+            & .left button {
+                font-size: 14px;
+                margin: 0 5px;
+            }
+            & .right .moreTag {
+                font-size: 14px;
+                padding: 0 5px;
+                margin-right: 10px;
+            }
+        }
+    }
+
+    .issueListBody {
+        & * {
+            font-size: 12px;
+        }
+        & .iconbox {
+            margin-right: 10px;
+    
+            & img {
+                height: 20px;
+            }
+            & p {
+                white-space: nowrap;
+            }
+        }
+        & .userbox img {
+            margin-top: 5px;
+        }
+        & .item {
+            padding: 10px 5px;
+        }
+        & span {
+            font-size: 12px !important;
+        }
+    
+        & .textbox .text2 {
+            font-size: 12px;
+            padding-left: 0;
+        }
+    }
+
+ 
+
+}
