+++ src/scss/component/style.scss
@@ -21,6 +21,7 @@
 @import "postitem.scss";
 @import "serviceitem.scss";
 @import "issuecard.scss";
+@import "issuelist.scss";
 
 // ide-sidebar-github
 @import "gitsystem.scss";
