+++ src/scss/project/issue.scss
@@ -9,158 +9,6 @@
 
     & .issuebox {
         position: relative;
-
-        & .titlebox {
-            @include flexbox(space-between, center);
-            height: 52px;
-            border: 1px solid #858585;
-            border-radius: 5px;
-            padding: 15px 20px;
-            background: #f2f2f2;
-            margin-bottom: 25px;
-
-            & .left {
-                position: relative;
-
-                & .btn_objective {
-                    font-size: 18px;
-                    color: #928a76;
-                    font-weight: bold;
-                }
-
-                & .objectivePopup {
-                    z-index: 1;
-                    position: absolute;
-                    top: 35px;
-                    left: -21px;
-                    width: 320px;
-                    height: 230px;
-                    border: 1px solid #1c2239;
-                    border-radius: 5px;
-                    background: #f2f2f2;
-                    padding: 5px 10px;
-
-                    & .searchbox {
-                        position: relative;
-
-                        & input {
-                            border: 1px solid #1c2239;
-                            border-radius: 5px;
-                            height: 23px;
-                            width: 100%;
-                        }
-
-                        & .btn_search {
-                            position: absolute;
-                            right: 5px;
-                            top: 3px;
-                        }
-                    }
-
-                    & h5 {
-                        font-size: 12px;
-                        color: #928a76;
-                        font-weight: bold;
-                        margin-top: 10px;
-                        margin-bottom: 5px;
-                    }
-
-                    & .items {
-                        position: relative;
-                        background: #fff;
-                        border-radius: 5px;
-                        padding: 5px;
-                        border: 1px solid #1c2239;
-                        height: 120px;
-                        overflow: auto;
-
-                        & button {
-                            position: relative;
-                            display: flex;
-                            justify-content: flex-start;
-                            align-items: center;
-                            width: 100%;
-                            text-align: left;
-                            padding-left: 20px;
-                            padding-bottom: 5px;
-                            font-weight: bold;
-                            font-size: 12px;
-
-                            & img {
-                                position: absolute;
-                                left: 0;
-                                top: -3px;
-                            }
-                        }
-                    }
-
-                    & .btnbox {
-                        @include flexbox(center, center);
-                        text-align: center;
-                        margin-top: 15px;
-
-                        & > * {
-                            margin: 0 20px;
-                            font-size: 12px;
-                        }
-
-                        & .btn_confirm {
-                            width: 105px;
-                            height: 28px;
-                            color: #fff;
-                            border-radius: 5px;
-                            background: #1c2239;
-                            font-weight: bold;
-                        }
-                        & .btn_del {
-                            color: #f95656;
-                            font-weight: bold;
-                        }
-                    }
-                } /* //objectivePopup */
-            }
-
-            & .right {
-                @include flexbox(space-between, center);
-
-                & > * {
-                    margin-left: 20px;
-                }
-
-                & .search {
-                    position: relative;
-                    width: 345px;
-                    height: 32px;
-
-                    & input {
-                        width: 100%;
-                        height: 100%;
-                        border-radius: 5px;
-                        background: #f2f2f2;
-                        border: 1px solid #858585;
-                        padding-right: 30px;
-                        color: #928a76;
-                        font-size: 14px;
-
-                        &::placeholder {
-                            color: #928a76;
-                            font-size: 14px;
-                            letter-spacing: 1px;
-                        }
-                    }
-
-                    & button {
-                        position: absolute;
-                        right: 7px;
-                        top: 7px;
-
-                        & img {
-                            height: 16px;
-                        }
-                    }
-                }
-            }
-        }
     }
 
     & .contentbox {
@@ -186,7 +34,7 @@
             }
         }
 
-        & .content {
+        & .cardContent {
             @include flexbox(space-between, flex-start);
             height: 100%;
             width: calc(100% - 100px);
